requirements.txt
streamlit
customtkinter
python-dotenv
requests
fpdf
pyttsx3
# optional: whisper (for voice input)